Changing the structure
Too many iterations of arrayfun - it was running once, and then running
again for every value. Instead, decreased the running time, but chaining
tasks together. Still working on it.

Consolidated the Sigmoid function, so now, it will return Sigmoid and
SigmoidPrime at the same time, since it is using the same initial value
to do the calculations. Also, the SigmoidPrime function was calling
Sigmoid, and therefore redundant.